Issue #SB-15476 chore: sync-assessment-events to buffer in-memory before sync
