+++ src/course/def/course-service.ts
@@ -14,6 +14,8 @@ import {Course} from './course';
 import {UnenrollCourseRequest} from './unenrollCourseRequest';
 import {DownloadCertificateRequest} from './download-certificate-request';
 import {DownloadCertificateResponse} from './download-certificate-response';
+import {SunbirdTelemetry} from '../../telemetry';
+import Telemetry = SunbirdTelemetry.Telemetry;
 
 export interface CourseService {
     getBatchDetails(request: CourseBatchDetailsRequest): Observable<Batch>;
@@ -33,6 +35,9 @@ export interface CourseService {
     downloadCurrentProfileCourseCertificate(
         downloadCertificateRequest: DownloadCertificateRequest
     ): Observable<DownloadCertificateResponse>;
+
+    /** @internal */
+    captureAssessmentEvent(capture: {event: Telemetry, courseContext: any});
 
     syncAssessmentEvents(): Observable<undefined>;
